Refactor before implementing file support in agents and UI
